Fixes bug when freezing text inside a multi-line entity
Not totally sure why it is happening, but I think that we were thawing the strings, but because of the way the transformers were arranged
the unthawing happened on the wrong piece of data.  And then we would discard the frozen-strings.
This does mean that frozen-strings could grow somewhat un-bounded, but this is markdown and so a few k (at worst?) of data should be ok.